+++ Frontend/src/assets/main.css
@@ -20,6 +20,113 @@
     font-weight: bold;
     font-style: bold;
 }
+/*********************entête*****************************/
+
+header{
+  background-color: #fff;
+}
+.second-bar{
+  background-color: #00367d;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 10px;
+
+  border: 0 0 0 5px solid #000;
+  box-shadow: 0px 8px 10px -5px rgba(0, 0, 0, 0.75);
+}
 .success{
     color: green;
+}
+
+/*********************tableau*****************************/
+
+table {
+  width: 100%;
+  border-collapse: collapse;
+  margin-top: 20px;
+  color: #36a5dd;
+}
+
+thead {
+  background-color: #36a5dd;
+  color: white;
+}
+
+th, td {
+  border: 1px solid #ddd;
+  padding: 8px;
+  text-align: left;
+}
+
+tr:hover {
+  background-color: #00367d;
+}
+
+/*********************formulaire*****************************/
+
+form {
+  max-width: 400px;
+  margin: 0 auto;
+}
+.form-group {
+  margin-bottom: 15px;
+}
+.input {
+  font-size: 13px; 
+}
+label {
+  display: block;
+  margin-bottom: 5px;
+  color: #00367d;
+}
+
+input {
+  width: 100%;
+  padding: 8px;
+  box-sizing: border-box;
+}
+
+.button, button {
+  background-color: #00367d;
+  color: #ffffff;
+  padding: 10px 20px;
+  border: none;
+  cursor: pointer;
+  border-radius: 5px;
+  transition: background-color 0.3s ease;
+  font-size: 15px; 
+}
+.button {
+  display: flex;
+  justify-content: center; 
+  align-items: center; 
+}
+
+button:hover {
+  background-color: #36a5dd;
+}
+
+.button:active {
+  background-color: #002855;
+}
+
+a {
+  text-decoration: none; 
+  color: white;          
+  cursor: pointer;       
+}
+
+.text{
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-top: 15vh;
+  text-align: center;
+}
+
+.mytext{
+  color: #00367d;
+  font-family: 'Poppins-Regular', sans-serif;
+  font-size: 13px; 
 }
