--- Frontend/src/components/enseignant/Ajouter_enseignant.vue
@@ -1,14 +1,0 @@
-<script setup lang="ts">
-
-</script>
-
-<template>
-  <div id="app" class="container-fluid" >
-    Hey
-  </div>
-</template>
-
-<style scoped>
-
-</style>
-
