+++ Frontend/src/views/Admin/StudentView/StudentAddView.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+import Header from '../../../components/Admin/Header.vue'
+import Add from '../../../components/Admin/Student/StudentAdd.vue'
+import Footer from '../../../components/Footer.vue'
+</script>
+
+<template>
+  <main>
+    <Header />
+    <Add />
+    <Footer />
+  </main>
+</template>
