+++ Frontend/src/views/Admin/StudentView/StudentListView.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+import Header from '../../../components/Admin/Header.vue'
+import List from '../../../components/Admin/Student/StudentList.vue'
+import Footer from '../../../components/Footer.vue'
+</script>
+
+<template>
+  <main>    
+    <Header />
+    <List />
+    <Footer />
+  </main>
+</template>
