+++ Frontend/src/views/Admin/StudentView/StudentMDPView.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+import Header from '../../../components/Admin/Header.vue'
+import Password from '../../../components/Admin/Student/StudentMDP.vue'
+import Footer from '../../../components/Footer.vue'
+</script>
+
+<template>
+  <main>
+    <Header />
+    <Password />
+    <Footer />
+  </main>
+</template>
