+++ Frontend/src/views/Admin/StudentView/StudentSupView.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+import Header from '../../../components/Admin/Header.vue'
+import Sup from '../../../components/Admin/Student/StudentSup.vue'
+import Footer from '../../../components/Footer.vue'
+</script>
+
+<template>
+  <main>
+    <Header />
+    <Sup />
+    <Footer />
+  </main>
+</template>
