+++ Frontend/src/views/Admin/TeacherView/TeacherAddView.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+import Header from '../../../components/Admin/Header.vue'
+import Add from '../../../components/Admin/Teacher/TeacherAdd.vue'
+import Footer from '../../../components/Footer.vue'
+</script>
+
+<template>
+  <main>
+    <Header />
+    <Add />
+    <Footer />
+  </main>
+</template>
