+++ Frontend/src/views/Admin/TeacherView/TeacherMDPView.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+import Header from '../../../components/Admin/Header.vue'
+import Password from '../../../components/Admin/Teacher/TeacherMDP.vue'
+import Footer from '../../../components/Footer.vue'
+</script>
+
+<template>
+  <main>
+    <Header />
+    <Password />
+    <Footer />
+  </main>
+</template>
