--- Frontend/src/views/AdministrateurView/AdminListView.vue
@@ -1,12 +1,0 @@
-<script setup lang="ts">
-import List from '../../components/Administrateur/AdminList.vue'
-import Footer from '../../components/Footer.vue'
-</script>
-
-<template>
-  <main>
-    
-    <List />
-    <Footer />
-  </main>
-</template>
