--- Frontend/src/views/AdministrateurView/AdminSupView.vue
@@ -1,12 +1,0 @@
-<script setup lang="ts">
-import Sup from '../../components/Administrateur/AdminSup.vue'
-import Footer from '../../components/Footer.vue'
-</script>
-
-<template>
-  <main>
-    
-    <Sup />
-    <Footer />
-  </main>
-</template>
