--- Frontend/src/views/AdministrateurView/AdminView.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-import Admin from '../../components/Administrateur/Admin.vue'
-import Footer from '../../components/Footer.vue'
-</script>
-
-<template>
-  <main>
-    <Admin />
-    <Footer />
-  </main>
-</template>
