+++ Frontend/src/views/Teach/NoteView/AddNoteView.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+import Header from '../../../components/Teach/Header.vue'
+import Add from '../../../components/Teach/Notes/AddNote.vue'
+import Footer from '../../../components/Footer.vue'
+</script>
+
+<template>
+  <main>
+    <Header />
+    <Add />
+    <Footer />
+  </main>
+</template>
