+++ Frontend/src/views/Teach/NoteView/ListView.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+import Header from '../../../components/Teach/Header.vue'
+import List from '../../../components/Teach/Notes/List.vue'
+import Footer from '../../../components/Footer.vue'
+</script>
+
+<template>
+  <main>    
+    <Header />
+    <List />
+    <Footer />
+  </main>
+</template>
