+++ Frontend/src/views/Teach/NoteView/ModNoteView.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+import Header from '../../../components/Teach/Header.vue'
+import Password from '../../../components/Teach/Notes/ModNote.vue'
+import Footer from '../../../components/Footer.vue'
+</script>
+
+<template>
+  <main>
+    <Header />
+    <Password />
+    <Footer />
+  </main>
+</template>
