--- Frontend/src/views/TeacherView.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-import Teacher from '../components/Teacher.vue'
-import Footer from '../components/Footer.vue'
-</script>
-
-<template>
-  <main>
-    <Teacher />
-    <Footer />
-  </main>
-</template>
