--- Frontend/src/views/enseignantView/AjouterEnseignantView.vue
@@ -1,10 +1,0 @@
-<script setup lang="ts">
-import Ajouter_enseignant from '../../components/enseignant/Ajouter_enseignant.vue'
-</script>
-
-<template>
-  <main>
-    
-    <Ajouter_enseignant />
-  </main>
-</template>
